Pull out path tests into new file
