+++ tests/cli-self-upd.rs
@@ -302,237 +302,9 @@ fn uninstall_doesnt_leave_gc_file() {
     })
 }
 
-#[cfg(unix)]
-fn install_adds_path_to_rc(rcfile: &str) {
-    setup(&|config| {
-        let my_rc = "foo\nbar\nbaz";
-        let rc = config.homedir.join(rcfile);
-        raw::write_file(&rc, my_rc).unwrap();
-        expect_ok(config, &["rustup-init", "-y"]);
-
-        let new_rc = fs::read_to_string(&rc).unwrap();
-        let addition = format!("source \"{}/env.sh\"", config.cargodir.display());
-        let expected = format!("{}\n{}\n", my_rc, addition);
-        assert_eq!(new_rc, expected);
-    });
-}
-
 #[test]
-#[cfg(unix)]
-fn install_adds_path_to_profile() {
-    install_adds_path_to_rc(".profile");
-}
-
-#[test]
-#[cfg(unix)]
-fn install_adds_path_to_bashrc() {
-    install_adds_path_to_rc(".bashrc");
-}
-
-#[test]
-#[cfg(unix)]
-fn install_adds_path_to_bash_profile() {
-    install_adds_path_to_rc(".bash_profile");
-}
-
-#[test]
-#[cfg(unix)]
-fn install_does_not_add_path_to_bash_profile_that_doesnt_exist() {
-    setup(&|config| {
-        let rc = config.homedir.join(".bash_profile");
-        expect_ok(config, &["rustup-init", "-y"]);
-
-        assert!(!rc.exists());
-    });
-}
-
-#[test]
-#[cfg(unix)]
-fn install_errors_when_rc_file_cannot_be_updated() {
-    setup(&|config| {
-        let rc = config.homedir.join(".bashrc");
-        fs::File::create(&rc).unwrap();
-        let mut perms = fs::metadata(&rc).unwrap().permissions();
-        perms.set_readonly(true);
-        fs::set_permissions(&rc, perms).unwrap();
-
-        expect_err(config, &["rustup-init", "-y"], "amend shell");
-    });
-}
-
-#[test]
-#[cfg(unix)]
-fn install_with_zsh_adds_path_to_zshenv() {
-    setup(&|config| {
-        let my_rc = "foo\nbar\nbaz";
-        let rc = config.homedir.join(".zshenv");
-        raw::write_file(&rc, my_rc).unwrap();
-
-        let mut cmd = clitools::cmd(config, "rustup-init", &["-y"]);
-        cmd.env("SHELL", "zsh");
-        assert!(cmd.output().unwrap().status.success());
-
-        let new_rc = fs::read_to_string(&rc).unwrap();
-        let addition = format!(r#"source "{}/env.sh""#, config.cargodir.display());
-        let expected = format!("{}\n{}\n", my_rc, addition);
-        assert_eq!(new_rc, expected);
-    });
-}
-
-#[test]
-#[cfg(unix)]
-fn install_with_zsh_adds_path_to_zdotdir_zshenv() {
-    setup(&|config| {
-        let zdotdir = tempfile::Builder::new()
-            .prefix("zdotdir")
-            .tempdir()
-            .unwrap();
-        let my_rc = "foo\nbar\nbaz";
-        let rc = zdotdir.path().join(".zshenv");
-        raw::write_file(&rc, my_rc).unwrap();
-
-        let mut cmd = clitools::cmd(config, "rustup-init", &["-y"]);
-        cmd.env("SHELL", "zsh");
-        cmd.env("ZDOTDIR", zdotdir.path());
-        assert!(cmd.output().unwrap().status.success());
-
-        let new_rc = fs::read_to_string(&rc).unwrap();
-        let addition = format!(r#"source "{}/env.sh""#, config.cargodir.display());
-        let expected = format!("{}\n{}\n", my_rc, addition);
-        assert_eq!(new_rc, expected);
-    });
-}
-
-#[test]
-#[cfg(unix)]
-fn install_adds_path_to_rcfile_just_once() {
-    setup(&|config| {
-        let my_profile = "foo\nbar\nbaz";
-        let profile = config.homedir.join(".profile");
-        raw::write_file(&profile, my_profile).unwrap();
-        expect_ok(config, &["rustup-init", "-y"]);
-        expect_ok(config, &["rustup-init", "-y"]);
-
-        let new_profile = fs::read_to_string(&profile).unwrap();
-        let addition = format!(r#"source "{}/env.sh""#, config.cargodir.display());
-        let expected = format!("{}\n{}\n", my_profile, addition);
-        assert_eq!(new_profile, expected);
-    });
-}
-
-#[cfg(unix)]
-fn uninstall_removes_path_from_rc(rcfile: &str) {
-    setup(&|config| {
-        let my_rc = "foo\nbar\nbaz";
-        let rc = config.homedir.join(rcfile);
-        raw::write_file(&rc, my_rc).unwrap();
-        expect_ok(config, &["rustup-init", "-y"]);
-        expect_ok(config, &["rustup", "self", "uninstall", "-y"]);
-
-        let new_rc = fs::read_to_string(&rc).unwrap();
-        assert_eq!(new_rc, my_rc);
-    });
-}
-
-#[test]
-#[cfg(unix)]
-fn uninstall_removes_path_from_profile() {
-    uninstall_removes_path_from_rc(".profile");
-}
-
-#[test]
-#[cfg(unix)]
-fn uninstall_removes_path_from_bashrc() {
-    uninstall_removes_path_from_rc(".bashrc");
-}
-
-#[test]
-#[cfg(unix)]
-fn uninstall_removes_path_from_bash_profile() {
-    uninstall_removes_path_from_rc(".bash_profile");
-}
-
-#[test]
-#[cfg(unix)]
-fn uninstall_removes_path_from_zshenv() {
-    uninstall_removes_path_from_rc(".zshenv");
-}
-
-#[test]
-#[cfg(unix)]
-fn uninstall_doesnt_touch_rc_files_that_dont_contain_cargo_home() {
-    setup(&|config| {
-        let my_rc = "foo\nbar\nbaz";
-        expect_ok(config, &["rustup-init", "-y"]);
-        expect_ok(config, &["rustup", "self", "uninstall", "-y"]);
-
-        let profile = config.homedir.join(".profile");
-        raw::write_file(&profile, my_rc).unwrap();
-
-        let profile = fs::read_to_string(&profile).unwrap();
-
-        assert_eq!(profile, my_rc);
-    });
-}
-
-// In the default case we want to write $HOME/.cargo/bin as the path,
-// not the full path.
-#[test]
-#[cfg(unix)]
-fn when_cargo_home_is_the_default_write_path_specially() {
-    setup(&|config| {
-        // Override the test harness so that cargo home looks like
-        // $HOME/.cargo by removing CARGO_HOME from the environment,
-        // otherwise the literal path will be written to the file.
-
-        let my_profile = "foo\nbar\nbaz";
-        let profile = config.homedir.join(".profile");
-        raw::write_file(&profile, my_profile).unwrap();
-        let mut cmd = clitools::cmd(config, "rustup-init", &["-y"]);
-        cmd.env_remove("CARGO_HOME");
-        assert!(cmd.output().unwrap().status.success());
-
-        let new_profile = fs::read_to_string(&profile).unwrap();
-        let expected = format!("{}\nsource \"$HOME/.cargo/env.sh\"\n", my_profile);
-        assert_eq!(new_profile, expected);
-
-        let mut cmd = clitools::cmd(config, "rustup", &["self", "uninstall", "-y"]);
-        cmd.env_remove("CARGO_HOME");
-        assert!(cmd.output().unwrap().status.success());
-
-        let new_profile = fs::read_to_string(&profile).unwrap();
-        assert_eq!(new_profile, my_profile);
-    });
-}
-
-#[test]
-#[cfg(windows)]
-/// Smoke test for end-to-end code connectivity of the installer path mgmt on windows.
-fn install_uninstall_affect_path() {
-    setup(&|config| {
-        let path = config.cargodir.join("bin").to_string_lossy().to_string();
-
-        expect_ok(config, &["rustup-init", "-y"]);
-        assert!(
-            get_path().unwrap().contains(&path),
-            format!("`{}` not in `{}`", get_path().unwrap(), &path)
-        );
-
-        expect_ok(config, &["rustup", "self", "uninstall", "-y"]);
-        assert!(!get_path().unwrap().contains(&path));
-    });
-}
-
-#[test]
-#[cfg(unix)]
-/// This covers both windows and unix no-path-modification cases due to the code structure.
-fn install_doesnt_modify_path_if_passed_no_modify_path() {
-    setup(&|config| {
-        let profile = config.homedir.join(".profile");
-        expect_ok(config, &["rustup-init", "-y", "--no-modify-path"]);
-        assert!(!profile.exists());
-    });
-}
+#[ignore]
+fn uninstall_stress_test() {}
 
 #[test]
 fn update_exact() {
@@ -918,24 +690,10 @@ fn reinstall_specifying_different_toolchain() {
 }
 
 #[test]
-#[cfg(unix)]
-fn produces_env_file_on_unix() {
+fn install_sets_up_stable() {
     setup(&|config| {
-        // Override the test harness so that cargo home looks like
-        // $HOME/.cargo by removing CARGO_HOME from the environment,
-        // otherwise the literal path will be written to the file.
-
-        let mut cmd = clitools::cmd(config, "rustup-init", &["-y"]);
-        cmd.env_remove("CARGO_HOME");
-        assert!(cmd.output().unwrap().status.success());
-        let envfile = config.homedir.join(".cargo/env.sh");
-        let envfile = fs::read_to_string(&envfile).unwrap();
-        let path_string = "export PATH=\"$HOME/.cargo/bin:${PATH}\"";
-        let (_, envfile_export) = envfile.split_at(match envfile.find("export PATH") {
-            Some(idx) => idx,
-            None => 0,
-        });
-        assert_eq!(&envfile_export[..path_string.len()], path_string);
+        expect_ok(config, &["rustup-init", "-y"]);
+        expect_stdout_ok(config, &["rustc", "--version"], "hash-stable-1.1.0");
     });
 }
